+++ crates/biome_js_formatter/src/context.rs
@@ -4,8 +4,8 @@ use crate::comments::{FormatJsLeadingComment, JsCommentStyle, JsComments};
 use biome_deserialize_macros::{Deserializable, Merge};
 use biome_formatter::printer::PrinterOptions;
 use biome_formatter::{
-    CstFormatContext, FormatContext, FormatElement, FormatOptions, IndentStyle, IndentWidth,
-    LineEnding, LineWidth, QuoteStyle, TransformSourceMap,
+    AttributePosition, CstFormatContext, FormatContext, FormatElement, FormatOptions, IndentStyle,
+    IndentWidth, LineEnding, LineWidth, QuoteStyle, TransformSourceMap,
 };
 use biome_js_syntax::{AnyJsFunctionBody, JsFileSource, JsLanguage};
 use std::fmt;
@@ -170,8 +170,8 @@ pub struct JsFormatOptions {
     /// Information related to the current file
     source_type: JsFileSource,
 
-    /// Enforce single attribute per line in HTML, Vue and JSX. Defaults to false
-    single_attribute_per_line: SingleAttributePerLine,
+    /// Attribute position style. By default auto.
+    attribute_position: AttributePosition,
 }
 
 impl JsFormatOptions {
@@ -190,7 +190,7 @@ impl JsFormatOptions {
             arrow_parentheses: ArrowParentheses::default(),
             bracket_spacing: BracketSpacing::default(),
             bracket_same_line: BracketSameLine::default(),
-            single_attribute_per_line: SingleAttributePerLine::default(),
+            attribute_position: AttributePosition::default(),
         }
     }
 
@@ -254,11 +254,8 @@ impl JsFormatOptions {
         self
     }
 
-    pub fn with_single_attribute_per_line(
-        mut self,
-        single_attribute_per_line: SingleAttributePerLine,
-    ) -> Self {
-        self.single_attribute_per_line = single_attribute_per_line;
+    pub fn with_attribute_position(mut self, attribute_position: AttributePosition) -> Self {
+        self.attribute_position = attribute_position;
         self
     }
 
@@ -305,11 +302,8 @@ impl JsFormatOptions {
     pub fn set_trailing_comma(&mut self, trailing_comma: TrailingComma) {
         self.trailing_comma = trailing_comma;
     }
-    pub fn set_single_attribute_per_line(
-        &mut self,
-        single_attribute_per_line: SingleAttributePerLine,
-    ) {
-        self.single_attribute_per_line = single_attribute_per_line;
+    pub fn set_attribute_position(&mut self, attribute_position: AttributePosition) {
+        self.attribute_position = attribute_position;
     }
 
     pub fn set_semicolons(&mut self, semicolons: Semicolons) {
@@ -356,8 +350,8 @@ impl JsFormatOptions {
         self.indent_width.value().into()
     }
 
-    pub fn single_attribute_per_line(&self) -> SingleAttributePerLine {
-        self.single_attribute_per_line
+    pub fn attribute_position(&self) -> AttributePosition {
+        self.attribute_position
     }
 }
 
@@ -376,6 +370,10 @@ impl FormatOptions for JsFormatOptions {
 
     fn line_ending(&self) -> LineEnding {
         self.line_ending
+    }
+
+    fn attribute_position(&self) -> AttributePosition {
+        self.attribute_position
     }
 
     fn as_print_options(&self) -> PrinterOptions {
@@ -397,11 +395,7 @@ impl fmt::Display for JsFormatOptions {
         writeln!(f, "Arrow parentheses: {}", self.arrow_parentheses)?;
         writeln!(f, "Bracket spacing: {}", self.bracket_spacing.value())?;
         writeln!(f, "Bracket same line: {}", self.bracket_same_line.value())?;
-        writeln!(
-            f,
-            "Single Attribute Per Line: {}",
-            self.single_attribute_per_line.value()
-        )
+        writeln!(f, "Attribute Position: {}", self.attribute_position)
     }
 }