+++ crates/biome_js_formatter/src/jsx/lists/attribute_list.rs
@@ -1,5 +1,5 @@
 use crate::prelude::*;
-use biome_formatter::FormatContext;
+use biome_formatter::AttributePosition;
 
 use biome_js_syntax::JsxAttributeList;
 
@@ -10,7 +10,7 @@ impl FormatRule<JsxAttributeList> for FormatJsxAttributeList {
     type Context = JsFormatContext;
 
     fn fmt(&self, node: &JsxAttributeList, f: &mut JsFormatter) -> FormatResult<()> {
-        let line_break = if f.context().options().single_attribute_per_line().value() {
+        let line_break = if f.options().attribute_position() == AttributePosition::Multiline {
             hard_line_break()
         } else {
             soft_line_break_or_space()