+++ website/src/playground/workers/biomeWorker.ts
@@ -1,12 +1,13 @@
 import init, {
-	type PartialConfiguration as Configuration,
 	DiagnosticPrinter,
+	type PartialConfiguration as Configuration,
 	type RomePath as BiomePath,
 	type RuleCategories,
 	Workspace,
 } from "@biomejs/wasm-web";
 import {
 	ArrowParentheses,
+	AttributePosition,
 	type BiomeOutput,
 	IndentStyle,
 	LintRules,
@@ -16,7 +17,7 @@ import {
 	QuoteStyle,
 	Semicolons,
 } from "../types";
-import { isJsonFilename } from "../utils";
+import {isJsonFilename} from "../utils";
 
 let workspace: Workspace | null = null;
 let fileCounter = 0;
@@ -78,6 +79,7 @@ self.addEventListener("message", async (e) => {
 				importSortingEnabled,
 				unsafeParameterDecoratorsEnabled,
 				allowComments,
+				attributePosition
 			} = e.data.settings as PlaygroundSettings;
 
 			configuration = {
@@ -87,6 +89,7 @@ self.addEventListener("message", async (e) => {
 					lineWidth: lineWidth,
 					indentStyle: indentStyle === IndentStyle.Tab ? "tab" : "space",
 					indentWidth,
+					attributePosition: attributePosition === AttributePosition.Auto ? 'auto' : 'multiline',
 				},
 
 				linter: {
@@ -115,6 +118,7 @@ self.addEventListener("message", async (e) => {
 								: "asNeeded",
 						bracketSpacing,
 						bracketSameLine,
+						attributePosition: attributePosition === AttributePosition.Auto ? "auto" : "multiline",
 					},
 					parser: {
 						unsafeParameterDecoratorsEnabled,