+++ website/src/playground/workers/prettierWorker.ts
@@ -5,6 +5,8 @@ import parserBabel from "prettier/esm/parser-babel.mjs";
 import pluginEstree from "prettier/plugins/estree.mjs";
 import {
 	ArrowParentheses,
+	AttributePosition,
+	defaultPlaygroundState,
 	IndentStyle,
 	type PlaygroundSettings,
 	type PrettierOptions,
@@ -13,9 +15,8 @@ import {
 	QuoteStyle,
 	Semicolons,
 	type TrailingComma,
-	defaultPlaygroundState,
 } from "../types";
-import { isJsonFilename, isTypeScriptFilename } from "../utils";
+import {isJsonFilename, isTypeScriptFilename} from "../utils";
 
 let settings = defaultPlaygroundState.settings;
 
@@ -39,6 +40,7 @@ self.addEventListener("message", async (e) => {
 				arrowParentheses,
 				bracketSpacing,
 				bracketSameLine,
+				attributePosition
 			} = settings;
 			const code = e.data.code as string;
 			const filename = e.data.filename as string;
@@ -56,6 +58,7 @@ self.addEventListener("message", async (e) => {
 				arrowParentheses,
 				bracketSpacing,
 				bracketSameLine,
+				singleAttributePerLine: attributePosition === AttributePosition.Multiline,
 			});
 
 			self.postMessage({
@@ -87,6 +90,7 @@ async function formatWithPrettier(
 		arrowParentheses: ArrowParentheses;
 		bracketSpacing: boolean;
 		bracketSameLine: boolean;
+		singleAttributePerLine?: boolean
 	},
 ): Promise<PrettierOutput> {
 	try {
@@ -108,6 +112,7 @@ async function formatWithPrettier(
 					: "avoid",
 			bracketSpacing: options.bracketSpacing,
 			bracketSameLine: options.bracketSameLine,
+			singleAttributePerLine: options.singleAttributePerLine ?? false,
 		};
 
 		// @ts-expect-error